Move interface SearchParams to types
