+++ packages/types/src/sdk/index.ts
@@ -19,3 +19,4 @@ export * from "./user"
 export * from "./cli"
 export * from "./websocket"
 export * from "./permissions"
+export * from "./row"
